requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
playwright==1.41.0
playwright-stealth==1.0.6
httpx==0.24.1
httpx-socks==0.7.7
python-multipart==0.0.6
jinja2==3.1.3
aiofiles==23.2.1
python-dateutil==2.8.2
socksio==1.0.0
python-dotenv==1.0.1
nest-asyncio==1.6.0
psutil==5.9.8
prometheus-fastapi-instrumentator==7.0.0
